- fix `difficulty_override` type `float→int`; add `pre_sense` test scaffold
~`pre_sense.py` Update difficulty_override type from float to int for consistency
^`dify-ky-pre_sense_test.py` Add initial test file for pre_sense node in Kaye Chat Dify app
?`dify-ky-post_start_test.py` Add FIXME comment to remove body reference in test file

+++ dify_studio/kaye_chat/nodes/sense/pre_sense.py
@@ -14,7 +14,7 @@ BODY_DIFF_KEY = "difficulty_override"
 
 
 # Entry Point  #################################################################
-def main(current_role: str, difficulty_override: float, query: str):
+def main(current_role: str, difficulty_override: int, query: str):
     # body  --------------------------------------------------------------------
     getter_body = json.dumps(
         {BODY_ROLE_KEY: current_role, BODY_DIFF_KEY: difficulty_override}
